- deleted scripts - migrated get_gr to its own file
