+++ ezfit/get_gr.py
@@ -1,0 +1,39 @@
+def get_gr(recipe):
+    """
+    Get the gr of a recipe and for each phase contribution
+    returns:
+    - r: list of floats
+    - gobs: list of floats
+    - gcalc: list of floats
+    - gdiff: list of floats
+    - baseline: float 
+    - gr_composition: dict of list of floats
+    """
+    def remove_consecutive_duplicates(string, char):
+        indices = [m.start() for m in re.finditer(char * 2, string)]
+        if indices:
+            for i in indices:
+                string = string[:i] + string[i+1:]
+            return remove_consecutive_duplicates(string, char)
+        else:
+            return string
+
+    equation = recipe.PDF.getEquation()        
+    for char in ['\)', '\(']:
+        equation = (remove_consecutive_duplicates(equation, char))
+    
+
+    prof = recipe._contributions['PDF'].profile
+    r = prof.x
+    gobs = prof.y
+    gcalc = recipe._contributions['PDF'].evaluate()
+    baseline = 1.35 * gobs.min()
+    gdiff = gobs - gcalc
+
+
+    gr_composition = {}
+    for eq in equation.split(' + '):
+        gr = recipe.PDF.evaluateEquation(eq[1:])
+        gr_composition[eq[1:]] = gr
+
+    return r, gobs, gcalc, gdiff, baseline, gr_composition
